Fix import path for FileOperationException and rename exception file
